refactor: split package class into subclasses
